Move some types and traits into data_model
